LPS-152441 Revamp portal-url-builder API
The existing API is a good step towards giving clear semantics to each URL
that we compose in the portal, server side, but it was still lacking enough
detail to control caching.

Also, some details related to pathProxy and CDN were misleading.

With this commit I'm creating one XXXPortalURLBuilder interface per type of
resource (we may add more in the future) and some cross-functional facet
interfaces that tell us what features can be controlled per URL type (eg:
cache, CDN, ...).

Hopefully this will be enough to base future URL compositions on this API
and route all of them through it.