LPS-152441 Implement the new portal-url-builder API (CacheHelper)
This class is the place where control of caching parameters like `t` or the
newly added `mac` parameter takes place.

The idea behind this class is to make resource update detection automatic
so that we can change the URL (using the `t` or `mac` parameters) to make
browsers invalidate their caches no matter what the HTTP headers say.

I'm implementing a cache of digests for static resources that is
invalidated as soon as any Bundle is reloaded. This is so that the server
only digests every single resource once per server restart or bundle
update.

There are more strategies possible (like adding a `t` with the last server
restart time) which make computing digests unneeded but create more
cache invalidation false positives in the browser.

The important thing about this API is that all these strategies are hidden
inside the API so we may change them in the future, make them configurable,
smart, etc. without any need to change the consumers' code.