LPS-152366 Releasing the desired priority position when exists a segment with that priority set.
To prevent duplicate entry index exception we need to update the priorities of the segments above or below of the target one, in order to be able later on to update the target segment using the desired priority.  This process should be execute in the proper order that grants that no priority could be duplicated.

Because of we need to keep the order during the process is required to prioritize all them. It could cause that the target segment will be updated too during this process, for example if its current priority if major than zero and we are decreasing its priority or is lower than zero and we are increasing its priority.