LPS-152441 Fix RTL support
We need to add some parameters to CSS and JS requests. I've copied them
from PortalImpl.getStaticResourceURL to break the dependency with legacy
code, given that each time we use AbsoluteURLBuilder it's because we are
either:

  a. Creating a new feature
  b. Refactoring some legacy code

So I think it's safe to make this assumption to force us to review
refactors and, at the same time, not to continue dragging a huge list of
parameters simply because they were there before.