LPS-164251 Don't render type=module for <script> nodes that don't use ES imports
This is because the es-module-shim we are using to polyfill import maps in some
browsers causes <script type=module> nodes to be executed twice under some
circumstances.

The es-module-shim expects all <script type=module> nodes to fail loading
because of the browser lack of support for bare imports. However, unless the first
import of ALL <script type=module> nodes is a bare import, that condition won't
hold because the browser can potentially execute some code before failing and
thus lead to double execution of that code when the shim kicks in.

To avoid that we won't render type=module unless necessary, for now.

In following PRs we may need to make this logic more complex because we don't use
bare imports internally, which means that all <script type=module> nodes we render
will be loaded successfully by the browser, leading to a double execution again
(maybe).

We will check if that happens and how to solve in a following PR, as I said.