LPS-165857 Add loading indicator when fetching menu items
In situations where the internet is slow or there is a race condition, opening the menu may render the menu blank because the request to fetch the Menu items is still in progress.

To make this more accessible for screen readers as well, add the appropriate loading indicator with semantic rules to inform the user that items are loading when the request is finish, moves the focus to the first menu item.